+++ projects/layer-control/src/lib/layer-control/layer-control.component.spec.ts
@@ -8,10 +8,11 @@ import { LayerentryComponent } from '../layerentry/layerentry.component';
 import { ClarityModule } from '@clr/angular';
 import { FormsModule } from '@angular/forms';
 // import { Layer, LayerGroup } from '@dlr-eoc/services-layers';
-import { ObjTypePipe } from '../utils/obj-type.pipe';
+import { ObjTypePipe, ItemsFilterPipe } from '../utils/obj-type.pipe';
 import { LayersService } from '@dlr-eoc/services-layers';
 import { MapStateService } from '@dlr-eoc/services-map-state';
 import { ReversePipe } from '../utils/array-reverse.pipe';
+import { DragDropModule } from '@angular/cdk/drag-drop';
 
 describe('LayerControlComponent', () => {
   let component: LayerControlComponent;
@@ -21,8 +22,8 @@ describe('LayerControlComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      imports: [ClarityModule, FormsModule],
-      declarations: [LayerControlComponent, LayerentryGroupComponent, LayerentryComponent, ObjTypePipe, ReversePipe],
+      imports: [ClarityModule, FormsModule, DragDropModule],
+      declarations: [LayerControlComponent, LayerentryGroupComponent, LayerentryComponent, ObjTypePipe, ReversePipe, ItemsFilterPipe],
       providers: [LayersService, MapStateService]
     }).compileComponents();
   }));