+++ projects/layer-control/src/lib/layerentry-group/layerentry-group.component.ts
@@ -37,6 +37,7 @@ export class LayerentryGroupComponent implements OnInit {
   public canZoomToGroup = false;
 
   public showInfo = false;
+  public showAction = true;
 
   constructor() { }
 
@@ -44,6 +45,37 @@ export class LayerentryGroupComponent implements OnInit {
     if (this.group.bbox && this.group.bbox.length >= 4) {
       this.canZoomToGroup = true;
     }
+
+    if (!this.group?.action) {
+      this.showAction = false;
+    }
+  }
+
+  /**
+   * obj: {any| IDynamicComponent}
+   */
+  checkIsComponentItem(group: LayerGroup, compProp: string) {
+    const obj = group[compProp];
+    let isComp = false;
+    if (obj && typeof obj === 'object') {
+      if ('component' in obj) {
+        if (!obj.inputs) {
+          const groupClone = Object.assign({}, group);
+          if (groupClone && groupClone[compProp]) {
+            delete groupClone[compProp];
+          }
+          obj.inputs = { group: groupClone };
+        } else if (obj.inputs && !obj.inputs.group) {
+          const groupClone = Object.assign({}, group);
+          if (groupClone && groupClone[compProp]) {
+            delete groupClone[compProp];
+          }
+          obj.inputs = Object.assign({ group: groupClone }, obj.inputs);
+        }
+        isComp = true;
+      }
+    }
+    return isComp;
   }
 
   checkBaselayer(group: LayerGroup) {
@@ -52,6 +84,11 @@ export class LayerentryGroupComponent implements OnInit {
     } else {
       return false;
     }
+  }
+
+  checkClassHide(layer: Layer) {
+    const hasHide = layer?.cssClass?.includes('hide') || false;
+    return !hasHide;
   }
 
   getLayerName(group: LayerGroup) {
@@ -99,6 +136,8 @@ export class LayerentryGroupComponent implements OnInit {
 
   showHideAllDetails() {
     this.openAllLayersProperties = !this.openAllLayersProperties;
+    this.showAction = this.openAllLayersProperties;
+    this.showInfo = this.openAllLayersProperties;
   }
 
   isFirst(group) {
@@ -112,10 +151,40 @@ export class LayerentryGroupComponent implements OnInit {
 
   // CDKDRagAndDrop -------------------------------------------------------------
   // https://material.angular.io/cdk/drag-drop/api
-  drop(event: CdkDragDrop<string[]>) {
-    const previousI = this.group.layers.length - event.previousIndex - 1;
-    const currentI = this.group.layers.length - event.currentIndex - 1;
-    moveItemInArray(this.group.layers, previousI, currentI);
+  drop(event: CdkDragDrop<Layer[]>) {
+    const groupLayers = this.group.layers;
+    const groupLeng = groupLayers.length;
+    const fiteredLayers = event.container.data; // filtered by [cdkDropListData]
+    const groupFiteredLeng = fiteredLayers.length;
+    let previousIFinal, newIFinal;
+
+    /**
+     * calc index with pipe reverse order
+     */
+    if (groupLeng === groupFiteredLeng) {
+      const previousIndex = groupLeng - event.previousIndex - 1;
+      const newIndex = groupLeng - event.currentIndex - 1;
+      previousIFinal = previousIndex;
+      newIFinal = newIndex;
+    } else {
+      /**
+       * If array is filtered get previousIndex by item.data and try to calculate ne index
+       * get layers for cdk indexes - 'connect' 'event.container.data' and the original not filtered data
+       */
+      const newLayer = fiteredLayers[event.currentIndex];
+      const previousIndex = groupLayers.findIndex(l => l.id === event.item.data.id);
+      let newIndex = groupLayers.findIndex(l => l.id === newLayer.id);
+
+      // Item is not moved
+      if (event.previousIndex === event.currentIndex) {
+        newIndex = previousIndex;
+      }
+
+      previousIFinal = previousIndex;
+      newIFinal = newIndex;
+    }
+
+    moveItemInArray(this.group.layers, previousIFinal, newIFinal);
     this.layersSvc.updateLayerGroup(this.group);
   }
 }