+++ projects/layer-control/src/lib/utils/obj-type.pipe.ts
@@ -9,3 +9,17 @@ export class ObjTypePipe implements PipeTransform {
     return observations.filter((o) => o.type === type);
   }
 }
+
+/**
+ * e.g. usage *ngFor="let item of items | itemsfilter: callbackfn"
+ */
+@Pipe({
+  name: 'itemsfilter', pure: false
+})
+export class ItemsFilterPipe implements PipeTransform {
+  transform(items: any[], callbackfn?: (value: any, index: number, array: any[]) => boolean): any[] {
+    /* TODO: check layerlist for purity console.log("reevaluated"); */
+    return items.filter(callbackfn);
+  }
+}
+